+++ Microsoft.Azure.Cosmos/src/Resource/ClientContextCore.cs
@@ -28,7 +28,6 @@ namespace Microsoft.Azure.Cosmos
         private readonly string userAgent;
         private readonly EncryptionProcessor encryptionProcessor;
         private readonly DekCache dekCache;
-        private readonly CleanupExpiredRawDekFromMemory cleanupExpiredRawDekFromMemory;
         private bool isDisposed = false;
 
         private ClientContextCore(
@@ -41,8 +40,7 @@ namespace Microsoft.Azure.Cosmos
             string userAgent,
             EncryptionProcessor encryptionProcessor,
             DekCache dekCache,
-            BatchAsyncContainerExecutorCache batchExecutorCache,
-            CleanupExpiredRawDekFromMemory cleanupExpiredRawDekFromMemory)
+            BatchAsyncContainerExecutorCache batchExecutorCache)
         {
             this.client = client;
             this.clientOptions = clientOptions;
@@ -54,7 +52,6 @@ namespace Microsoft.Azure.Cosmos
             this.encryptionProcessor = encryptionProcessor;
             this.dekCache = dekCache;
             this.batchExecutorCache = batchExecutorCache;
-            this.cleanupExpiredRawDekFromMemory = cleanupExpiredRawDekFromMemory;
         }
 
         internal static CosmosClientContext Create(
@@ -122,8 +119,6 @@ namespace Microsoft.Azure.Cosmos
 
             CosmosResponseFactory responseFactory = new CosmosResponseFactory(serializerCore);
 
-            CleanupExpiredRawDekFromMemory cleanupExpiredRawDekFromMemory = new CleanupExpiredRawDekFromMemory(startProcess: clientOptions.EncryptionKeyWrapProvider != null);
-
             return new ClientContextCore(
                 client: cosmosClient,
                 clientOptions: clientOptions,
@@ -133,9 +128,8 @@ namespace Microsoft.Azure.Cosmos
                 documentClient: documentClient,
                 userAgent: documentClient.ConnectionPolicy.UserAgentContainer.UserAgent,
                 encryptionProcessor: new EncryptionProcessor(),
-                dekCache: new DekCache(),
-                batchExecutorCache: new BatchAsyncContainerExecutorCache(),
-                cleanupExpiredRawDekFromMemory: cleanupExpiredRawDekFromMemory);
+                dekCache: new DekCache(dekPropertiesTimeToLive: null, startCleanupProcess: clientOptions.EncryptionKeyWrapProvider != null),
+                batchExecutorCache: new BatchAsyncContainerExecutorCache());
         }
 
         /// <summary>
@@ -158,8 +152,6 @@ namespace Microsoft.Azure.Cosmos
         internal override EncryptionProcessor EncryptionProcessor => this.ThrowIfDisposed(this.encryptionProcessor);
 
         internal override DekCache DekCache => this.ThrowIfDisposed(this.dekCache);
-
-        internal override CleanupExpiredRawDekFromMemory CleanupExpiredRawDekFromMemory => this.ThrowIfDisposed(this.cleanupExpiredRawDekFromMemory);
 
         /// <summary>
         /// Generates the URI link for the resource
@@ -401,7 +393,7 @@ namespace Microsoft.Azure.Cosmos
                 {
                     this.batchExecutorCache.Dispose();
                     this.DocumentClient.Dispose();
-                    this.cleanupExpiredRawDekFromMemory.Dispose();
+                    this.dekCache.CleanupExpiredRawDekFromMemory.Dispose();
                 }
 
                 this.isDisposed = true;