+++ Microsoft.Azure.Cosmos/src/Resource/CosmosClientContext.cs
@@ -41,8 +41,6 @@ namespace Microsoft.Azure.Cosmos
 
         internal abstract DekCache DekCache { get; }
 
-        internal abstract CleanupExpiredRawDekFromMemory CleanupExpiredRawDekFromMemory { get; }
-
         internal abstract BatchAsyncContainerExecutor GetExecutorForContainer(
             ContainerCore container);