+++ changelog.md
@@ -16,7 +16,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - [#1312](https://github.com/Azure/azure-cosmos-dotnet-v3/pull/1312) Fixed null reference when using default(PartitionKey)
 - [#1296](https://github.com/Azure/azure-cosmos-dotnet-v3/pull/1296) Decrypt the encrypted properties before returning query result
-- [#1331](https://github.com/Azure/azure-cosmos-dotnet-v3/pull/1331) Enable client encryption / decryption for transactional batch requests
+- [#1331](https://github.com/Azure/azure-cosmos-dotnet-v3/pull/1331) Add client encryption / decryption for transactional batch requests
+- [#1352](https://github.com/Azure/azure-cosmos-dotnet-v3/pull/1352) Add capability to delete raw DEK from memory after expiry time
 
 ## <a name="3.7.1-preview"/> [3.7.0-preview2](https://www.nuget.org/packages/Microsoft.Azure.Cosmos/3.7.1-preview) - 2020-03-30